LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/overflow-clip-margin-hit-testing-expected.txt


FAIL Ensure elements in overflow-clip-margin are returned from elementFromPoint assert_equals: in child and container overflow clip margin expected Element node <body><div id="container" style="width: 100px; height: 10... but got Element node <div id="child" style="width: 300px; height: 50px"></div>
